Converted to custom hook and simplified interval logic
